requirements.txt
absl-py==2.1.0
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cvlib==0.2.7
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.15.4
Flask==3.1.0
flatbuffers==23.5.26
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-genai==0.3.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.27.1
idna==3.6
imageio==2.35.1
imbalanced-learn==0.13.0
imblearn==0.0
importlib-metadata==6.11.0
imutils==0.5.4
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.13
keras==2.15.0
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.47
libclang==16.0.6
lightgbm==4.5.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.0
oauthlib==3.2.2
openai==1.58.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.1
parso==0.8.3
patsy==1.0.1
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.2.0
progressbar==2.5
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.6.3
Pympler==1.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.11.1
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.25.0
sympy==1.13.3
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
transformers==4.48.1
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.3.1
ultralytics==8.3.27
ultralytics-thop==2.0.10
uri-template==1.3.0
urllib3==2.2.0
validators==0.34.0
virtualenv==20.26.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==14.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.14.1
xgboost==2.1.3
yarl==1.9.4
zipp==3.21.0
